refactor(whatsapp inbox): address CodeFactor findings on the media change
CodeFactor flagged the new code where it had grown too dense, and one real
lint error. All three are in what this branch added.

- Extract the attachment and voice-recorder concerns out of ReplyBox into a
  useAttachment hook. The component had absorbed upload, validation and the
  whole MediaRecorder lifecycle on top of the text and template flows.
- Drop a useless spread when building the send options (unicorn/no-useless-spread)
  — the object literal held nothing but the spread.
- Split WatiMessageProvider.sendMedia: the media download and the response
  reading are now their own methods. The multipart path was already long and
  this change had made it longer.

Also closes a hole in the media gate: it only refused when the capability probe
had answered 'no', so a probe that was slow or failed with a network error left
it 'unknown' — and an attachment sent in that state would have been delivered as
caption-only text with the file dropped. Only 'yes' is permission now, checked
at send time rather than only when picking a file.

101 backend tests pass; tsc clean; the new hook lints clean and the three
touched files now sit below their pre-existing lint count.
